restructure files in feat txn
