+++ lib/features/transactions/presentation/presenters/contributors/onchain_section_contributor.dart
@@ -1,6 +1,6 @@
 import 'package:bb_mobile/features/transactions/domain/entities/transaction.dart';
 import 'package:bb_mobile/features/transactions/presentation/presenters/view_models/transaction_detail_view_model.dart';
-import 'package:bb_mobile/features/transactions/presentation/presenters/transaction_section_contributor.dart';
+import 'package:bb_mobile/features/transactions/presentation/presenters/contributors/transaction_section_contributor.dart';
 import 'package:bb_mobile/features/transactions/utils/tx_format.dart';
 
 /// Plain on-chain Bitcoin/Liquid sends and receives. Composes under payjoin
