+++ lib/features/transactions/presentation/presenters/contributors/payjoin_section_contributor.dart
@@ -1,7 +1,7 @@
 import 'package:bb_mobile/core/payjoin/domain/entity/payjoin.dart';
 import 'package:bb_mobile/features/transactions/domain/entities/transaction.dart';
 import 'package:bb_mobile/features/transactions/presentation/presenters/view_models/transaction_detail_view_model.dart';
-import 'package:bb_mobile/features/transactions/presentation/presenters/transaction_section_contributor.dart';
+import 'package:bb_mobile/features/transactions/presentation/presenters/contributors/transaction_section_contributor.dart';
 import 'package:bb_mobile/features/transactions/utils/tx_format.dart';
 
 /// Payjoin. Composes on top of [OnchainSectionContributor] once the payjoin is
