+++ lib/features/transactions/presentation/presenters/transaction_detail_view_model_builder.dart
@@ -1,6 +1,6 @@
 import 'package:bb_mobile/features/transactions/domain/entities/transaction.dart';
 import 'package:bb_mobile/features/transactions/presentation/presenters/view_models/transaction_detail_view_model.dart';
-import 'package:bb_mobile/features/transactions/presentation/presenters/transaction_section_contributor.dart';
+import 'package:bb_mobile/features/transactions/presentation/presenters/contributors/transaction_section_contributor.dart';
 
 /// Assembles a [TransactionDetailViewModel] by running every registered
 /// [TransactionSectionContributor] that applies to the transaction.
