+++ lib/features/transactions/transactions_locator.dart
@@ -27,10 +27,10 @@ import 'package:bb_mobile/features/transactions/application/usecases/get_transac
 import 'package:bb_mobile/features/transactions/presentation/blocs/export/export_transactions_cubit.dart';
 import 'package:bb_mobile/features/transactions/presentation/blocs/transaction_details/transaction_details_cubit.dart';
 import 'package:bb_mobile/features/transactions/presentation/blocs/transactions_cubit.dart';
-import 'package:bb_mobile/features/transactions/presentation/presenters/onchain_section_contributor.dart';
-import 'package:bb_mobile/features/transactions/presentation/presenters/order_section_contributor.dart';
-import 'package:bb_mobile/features/transactions/presentation/presenters/payjoin_section_contributor.dart';
-import 'package:bb_mobile/features/transactions/presentation/presenters/swap_section_contributor.dart';
+import 'package:bb_mobile/features/transactions/presentation/presenters/contributors/onchain_section_contributor.dart';
+import 'package:bb_mobile/features/transactions/presentation/presenters/contributors/order_section_contributor.dart';
+import 'package:bb_mobile/features/transactions/presentation/presenters/contributors/payjoin_section_contributor.dart';
+import 'package:bb_mobile/features/transactions/presentation/presenters/contributors/swap_section_contributor.dart';
 import 'package:bb_mobile/features/transactions/presentation/presenters/transaction_detail_view_model_builder.dart';
 import 'package:get_it/get_it.dart';
 
@@ -110,9 +110,7 @@ class TransactionsLocator {
         saver: locator<TransactionExportSaver>(),
       ),
     );
-    // Registry of section contributors. Adding support for a new mechanism
-    // (e.g. silent payments, another L2) means adding a contributor here —
-    // nothing else changes (Open/Closed).
+    // Registry of section contributors
     locator.registerLazySingleton<TransactionDetailViewModelBuilder>(
       () => const TransactionDetailViewModelBuilder([
         OnchainSectionContributor(),
